Endpoint para autenticar users (#20)
* Added the function loginUser to the users router

* Created the loginUser function to the users controller

* Created the route /auth

* Added the loginUser function to the auth router

* Deleted the loginUser function from the users router

* Deleted the loginUser funtion from the users controller

* Created the loginUser function to the auth controller
