+++ routes/index.js
@@ -2,6 +2,7 @@ const express = require('express')
 const usersRouter = require('./users')
 const transactionsRouter = require('./transactions')
 const categoriesRouter = require('./categories')
+const authRouter = require('./auth')
 
 const router = express.Router()
 
@@ -9,5 +10,6 @@ const router = express.Router()
 router.use('/users', usersRouter)
 router.use('/transactions', transactionsRouter)
 router.use('/categories', categoriesRouter)
+router.use('/auth', authRouter)
 
 module.exports = router
